feat(#538): Fixes FireFox issue with tables and does some refactoring
