+++ assets/js/react/components/common/DocumentsTable/customHeaderCellRenders.js
