+++ assets/js/react/components/common/DocumentsTable/index.jsx
@@ -3,44 +3,13 @@ import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 import makeCustomCellRenders from './customCellRenders';
 import makeCustomHeaderCellRenders from './customHeaderCellRenders';
-import Table from '../../components/common/Table';
-import {
-  createColumns,
-  createDefaultHiddenColumns,
-} from '../../components/common/Table/TableHead';
-
-const defaultHiddenColumns = [
-  'author',
-  'authorComment',
-  'authors',
-  'cave',
-  'datePublication',
-  'dateValidation',
-  'descriptions',
-  'editor',
-  'entrance',
-  'files',
-  'id',
-  'identifier',
-  'identifierType',
-  'isValidated',
-  'library',
-  'license',
-  'massif',
-  'pages',
-  'parent',
-  'pathOld',
-  'publication',
-  'publicationFasciculeBBSOld',
-  'refBbs',
-  'reviewer',
-  'validationComment',
-  'validator',
-];
+import Table from '../Table';
+import { createColumns, createDefaultHiddenColumns } from '../Table/TableHead';
 
 const DocumentsTable = ({
   currentPage,
   documents,
+  defaultHiddenColumns,
   loading,
   openDetailedView,
   order,
@@ -63,11 +32,7 @@ const DocumentsTable = ({
   );
 
   useEffect(() => {
-    setColumns(
-      createColumns(documents, makeTranslation).sort(
-        (c1, c2) => c1.label.localeCompare(c2.label) > 0,
-      ),
-    );
+    setColumns(createColumns(documents, makeTranslation));
   }, [documents]);
 
   useEffect(() => {
@@ -108,6 +73,7 @@ DocumentsTable.propTypes = {
       title: PropTypes.string,
     }),
   ),
+  defaultHiddenColumns: PropTypes.arrayOf(PropTypes.string),
   loading: PropTypes.bool.isRequired,
   openDetailedView: PropTypes.func.isRequired,
   order: PropTypes.oneOf(['asc', 'desc']).isRequired,
